requirements.txt
streamlit==1.28.1
openai==0.27.8
python-dotenv==1.0.0